feat: Extract carousel to it's own component
