+++ app/components/ui/Dashboard.tsx
@@ -5,6 +5,7 @@ import { FormFrontend } from "@/app/schemas/formSchema";
 import { WhatWouldYouChoose } from "../Dashboard/Cards/WhatWouldYouChoose";
 import { WhatDifference } from "../Dashboard/Cards/WhatDifference";
 import { HowMuchFHCost } from "../Dashboard/Cards/HowMuchFHCost";
+import { Carousel } from "./Carousel";
 
 interface DashboardProps {
   processedData: Household;
@@ -14,27 +15,6 @@ interface DashboardProps {
 const Dashboard: React.FC<DashboardProps> = ({ inputData, processedData }) => {
   const scrollContainerRef = useRef<HTMLDivElement>(null);
   const [currentPage, setCurrentPage] = useState(0);
-  const totalPages = 5;
-
-  const handleNext = () => {
-    if (scrollContainerRef.current) {
-      scrollContainerRef.current.scrollBy({
-        top: window.innerHeight,
-        behavior: "smooth",
-      });
-    }
-    setCurrentPage(currentPage + 1);
-  };
-
-  const handlePrevious = () => {
-    if (scrollContainerRef.current) {
-      scrollContainerRef.current.scrollBy({
-        top: - window.innerHeight,
-        behavior: "smooth",
-      });
-    }
-    setCurrentPage(currentPage - 1);
-  };
 
   const handleScroll = () => {
     if (scrollContainerRef.current) {
@@ -55,18 +35,7 @@ const Dashboard: React.FC<DashboardProps> = ({ inputData, processedData }) => {
         ref={scrollContainerRef}
         onScroll={handleScroll}
       >
-        {currentPage > 0 && (
-          <div className="fixed top-4 right-4">
-            <button
-              onClick={handlePrevious}
-              className="px-4 py-2 bg-blue-500 text-white rounded hover:bg-blue-600"
-            >
-              Previous
-            </button>
-          </div>
-        )}
         <HowMuchFHCost data={processedData} />
-
         <GraphCard title="How much would it cost every month?">
           <span className="text-red-500">in theory less than Freehold</span>
         </GraphCard>
@@ -75,16 +44,10 @@ const Dashboard: React.FC<DashboardProps> = ({ inputData, processedData }) => {
         <WhatDifference />
         <WhatWouldYouChoose />
       </div>
-      {currentPage < totalPages && (
-        <div className="fixed bottom-4 right-4">
-          <button
-            onClick={handleNext}
-            className="px-4 py-2 bg-blue-500 text-white rounded hover:bg-blue-600"
-          >
-            Next
-          </button>
-        </div>
-      )}
+      <Carousel
+        scrollContainerRef={scrollContainerRef}
+        currentPage={currentPage}
+      />
     </div>
   );
 };
